Calculate Treatment Costs

This function calculates the treatment costs, , `Dinutuximab β` (GD2) and `Isotretinoin` (TT). including the costs of managing any associated adverse events, over a given time horizon.

Usage

calculate_treatment_costs(
  l_params,
  GD2_cycle_days = 35,
  TT_cycle_days = 30,
  Temo_cycle_days = 21,
  Iri_cycle_days = 21
)

Arguments

l_params

A list of model parameters including the list of parameters in the list l_psm_parameters in addition to:

time_horizon: The time horizon for the model in years.

cycle_length: The length of a model cycle measured in years.

disc_rate_costs: The annual discount rate for incurred costs.

disc_rate_qalys: The annual discount rate for accrued QALYs.

GD2_cycle_days

The number of days in a GD2 treatment cycle. Default value is `35` days.

TT_cycle_days

The number of days in a TT treatment cycle. Default value is `30` days.

Temo_cycle_days

The number of days in a Temozolomide treatment cycle. Default value is `21` days.

Iri_cycle_days

The number of days in a Irinotecan treatment cycle. Default value is `21` days.

Value

A matrix of treatment costs for the first year, with columns for GD2 and TT costs over the specified time points.

Examples

if (FALSE) { # \dontrun{
# Define model parameters
params <- c(
  time_horizon = 10,
  cycle_length = 1/12,
  disc_rate_costs = 0.035,
  disc_rate_qalys = 0.015,
  NeuroblastomaPSM::l_psm_parameters
)

# Calculate treatment costs
l_treatment_costs <- NeuroblastomaPSM::calculate_treatment_costs(
   l_params = params
)

View(l_treatment_costs)
} # }