Probabilistic Sensitivity Analysis (PSA) parameters

This list contains the parameters used for a probabilistic sensitivity analysis (PSA). Each parameter includes the distribution function and the necessary parameters for that distribution.

Usage

l_psa_parameters

Format

A list with the following components:

body_weight

A list with the following elements:

dist_func

Distribution function, here "rnorm".

dist_params

A list of parameters for the distribution:

mean

Mean of the normal distribution.

sd

Standard deviation of the normal distribution.

GD2_unit_price

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

TT_unit_price

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

prob_anemia

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_leukopenia

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_neutropenia

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_thrombocytopenia

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_fever

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_urticaria

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_immunotherapy_related_pain

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_diarrhea

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_hypotension

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_vomiting

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_infection

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_hypersensitivity_reaction

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

prob_capillary_leak

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

cost_anemia

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_leukopenia

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_neutropenia

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_thrombocytopenia

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_fever

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_urticaria

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_immunotherapy_related_pain

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_diarrhea

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_hypotension

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_vomiting

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_infection

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_hypersensitivity_reaction

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_capillary_leak

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_pain

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_laboratory

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

cost_hospitalisation

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

Temo_unit_price

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

Iri_unit_price

A list with the following elements:

dist_func

Distribution function, here "rgamma".

dist_params

A list of parameters for the distribution:

shape

Shape parameter of the gamma distribution.

scale

Scale parameter of the gamma distribution.

u_EFS

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.

u_PPS

A list with the following elements:

dist_func

Distribution function, here "rbeta".

dist_params

A list of parameters for the distribution:

alpha

Alpha parameter of the beta distribution.

beta

Beta parameter of the beta distribution.