Partitioned Survival Model (PSM) Parameters

This list contains the Partitioned Survival Model (PSM) parameters including various cost, probability, and utility values used in the model.

Usage

l_psm_parameters

Format

A list with the following components:

age

Starting age of simulated cohort, 4.

body_weight

Body weight in kilograms, 15.

smr_EFS

Standardized Mortality Ratio (SMR) for the Event Free Survival compared to the general population all-cause mortality, 5.6.

smr_PPS

Standardized Mortality Ratio (SMR) for the Post Progression Survival compared to the general population all-cause mortality, 10.64.

cure_threshold

The time in years at which those in the EFS are considered cured, 10. Beyond this time, the mortality hazard rates for the EFS and PPS is set by the `smr_EFS` and `smr_PPS` parameters, respectively.

GD2_unit_mg

Unit dose of Dinutuximab β (GD2) in milligrams, 20.

GD2_dose_days

Number of days on Dinutuximab β per cycle, 5.

GD2_unit_price

Unit price of Dinutuximab β, 8790.26.

TT_unit_mg

Unit dose of Isotretinoin (TT) in milligrams, 20.

TT_dose_days

Number of days for Isotretinoin dosing, 14.

TT_unit_price

Unit price of Isotretinoin, 0.24.

GD2_aEvents_names

Names of adverse events for GD2:

anemia

anemia

leukopenia

leukopenia

neutropenia

neutropenia

thrombocytopenia

thrombocytopenia

fever

Fever

urticaria

urticaria

immunotherapy_related_pain

immunotherapy related pain

diarrhea

Diarrhea

hypotension

hypotention

vomiting

Vomiting

infection

Infection

hypersensitivity_reaction

Hypersensitivity reaction

capillary_leak

Capillary leak

prob_anemia

Probability of anemia, 0.42

prob_leukopenia

Probability of leukopenia, 0.45

prob_neutropenia

Probability of neutropenia, 0.33

prob_thrombocytopenia

Probability of thrombocytopenia, 0.35

prob_fever

Probability of Fever, 0.14

prob_urticaria

Probability of urticaria, 0.05

prob_immunotherapy_related_pain

Probability of immunotherapy related pain, 0.16

prob_diarrhea

Probability of Diarrhea, 0.07

prob_hypotension

Probability of hypotention, 0.08

prob_vomiting

Probability of Vomiting, 0.06

prob_infection

Probability of Infection, 0.25

prob_hypersensitivity_reaction

Probability of Hypersensitivity reaction, 0.08

prob_capillary_leak

Probability of Capillary leak, 0.1

cost_anemia

Cost of anemia, 0.

cost_leukopenia

Cost of leukopenia, 0

cost_neutropenia

Cost of neutropenia, 0

cost_thrombocytopenia

Cost of thrombocytopenia, 0

cost_fever

Cost of Fever, 1.98

cost_urticaria

Cost of urticaria, 0

cost_immunotherapy_related_pain

Cost of immunotherapy related pain, 0

cost_diarrhea

Cost of Diarrhea, 3.2

cost_hypotension

Cost of hypotention, 0

cost_vomiting

Cost of Vomiting, 8.45

cost_infection

Cost of Infection, 222.33

cost_hypersensitivity_reaction

Cost of Hypersensitivity reaction, 4.13

cost_capillary_leak

Cost of Capillary leak, 306.37

cost_pain

Cost of pain, 6.36.

cost_laboratory

Cost of laboratory services, 141.45.

cost_hospitalisation

Cost of hospitalisation, 3763.62.

Temo_unit_mg

Unit dose of Temozolomide (Temo) in milligrams, 50.

Temo_dose_days

Number of days on Temozolomide per cycle, 5.

Temo_unit_price

Unit price of Temozolomide, 15.34.

Iri_unit_mg

Unit dose of Irinotecan (Iri) in milligrams, 40.

Iri_dose_days

Number of days on Irinotecan per cycle, 5.

Iri_unit_price

Unit price of Irinotecan, 4.4.

u_EFS

Utility value for Event-Free Survival (EFS), 0.84.

u_PPS

Utility value for Post-Progression Survival (PPS), 0.56.

df_life_table

dataframe representing the Life Table for Jordan:

age

Numeric vector representing the age groups from 0 to 100.

prob

Numeric vector representing the mortality rates.